Parse Ruby blocks correctly. Fixes #1
The parser used to continue reading (and expecting) Ruby lines
when a Ruby line ended in a backslash, e.g.

```haml
- a |
  b |
```

This behavior is incorrect when using Ruby block syntax, e.g.

```haml
- a do |b|
  .div
```

where the first HAML line after the Ruby line was read as
a continuation of the Ruby line, and the HAML parsing started
only at the second subsequent HAML line.

This misbehavior manifested in malformed indentation errors
when the subsequent HAML tree had more than one branch
at the root node.

We fix this by making the parser aware of Ruby block syntax.